Moved TLS 1.2 configuration to a more central location
The reason for this, is that the configuration is consistent. If you
called CPR Services and afterwards called something else, you would be
using TLS 1.2. But if you did not call CPR Services you would not use
TLS 1.2.
Now it will always use TLS 1.2, because it is "configured" in a central
location